add-cart1, view-cart1

+++ src/main/java/vn/com/devmaster/services/managematerial/projection/IViewProduct.java
@@ -5,6 +5,6 @@ public interface IViewProduct {
     String getName();
     Integer getQuantity();
     String getImage();
-    Integer getPrice();
+    Double getPrice();
 
 }
